src/diagram/requirements_3d.txt
matplotlib>=3.5.0
numpy>=1.21.0
numpy-stl>=3.0.0
